+++ cmd/frontend/internal/search/repos/repos_test.go
@@ -150,7 +150,7 @@ func TestRevisionValidation(t *testing.T) {
 		t.Run(tt.repoFilters[0], func(t *testing.T) {
 
 			op := Options{RepoFilters: tt.repoFilters}
-			resolved, err := Resolve(context.Background(), op)
+			resolved, err := ResolveRepositories(context.Background(), op)
 
 			if diff := cmp.Diff(tt.wantRepoRevs, resolved.RepoRevs); diff != "" {
 				t.Error(diff)
